% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_ic.R
\name{calcular_ic}
\alias{calcular_ic}
\title{Función calcular_ic}
\usage{
calcular_ic(x, y, confianza = 0.95, unidad = NULL, estadistico = "Z")
}
\arguments{
\item{x}{Variable numérica - diferencia de medias o categórica - diferencia de proporciones.}

\item{y}{Variable categórica de 2 niveles.}

\item{confianza}{Confianza de nuestro intervalo. - En decimales.}

\item{unidad}{Unidad de la variable numérica.- Opcional}

\item{estadistico}{Definir si usar "Z" o "T" en la formula del IC, dependiendo de la distribución asumida.}
}
\description{
Esta función nos permite calcular intervalos de confianza para diferencias de medias y proporciones.
}
\examples{
calcular_ic()

}
\keyword{calcular_ic}
\keyword{leacs}
