% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_tablas.R
\name{crear_tablas}
\alias{crear_tablas}
\title{Función crear_tablas}
\usage{
crear_tablas(data)
}
\arguments{
\item{data}{Base de datos que contiene las variables categóricas.}
}
\description{
Esta función nos permite crear tablas para todas las variables asignadas como factores en una base de datos.
}
\examples{
crear_tablas()

}
\keyword{crear_tablas}
\keyword{leacs}
