% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_tablas2.R
\name{crear_tablas2}
\alias{crear_tablas2}
\title{Función crear_tablas2}
\usage{
crear_tablas2(data, categorica_principal, margen = 2)
}
\arguments{
\item{data}{Base de datos que contiene las variables categóricas.}

\item{categorica_principal}{Variable categórica en la que se dividen las otras variables.}

\item{margen}{Elección de como se representan los porcentajes en la tabla 2x2 de proporciones..}
}
\description{
Esta función nos permite crear tablas 2x2 para todas las variables asignadas como factores en una base de datos en función de una variable categórica principal.
}
\examples{
crear_tablas2()

}
\keyword{crear_tablas2}
\keyword{leacs}
