% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describir.R
\name{describir}
\alias{describir}
\title{Función describir}
\usage{
describir(
  data,
  var.num,
  nombre.num = "",
  unidad = "",
  var.cat = NULL,
  nombre.cat = "",
  label = NULL,
  redondeo = 2,
  hist = TRUE,
  bins = NULL,
  colores = c("#457b9d", "#2a9d8f", "#2a9d9f"),
  box = TRUE,
  qq = TRUE,
  shapiro = TRUE,
  guardar = FALSE
)
}
\arguments{
\item{var.num}{Variable numérica a describir.}

\item{nombre.num}{Nombre de la variable que aparecera en outputs - opcional.}

\item{unidad}{Unidad de la variable que aparece en outputs - opcional.}

\item{var.cat}{Variable categórica a utilizar si hace un analisis bivariado.}

\item{nombre.cat}{Nombre de la variable categórica que aparece en outputs - opcional.}

\item{label}{Labels de los niveles de la variable - opcional.}

\item{redondeo}{Cantidad de decimales a usar en los redondeos. Default es 2.}

\item{hist}{Graficar o no un histograma. Default en TRUE.}

\item{bins}{Cantidad de bins del histograma. Por default hace una cuenta aproximada de bins óptimos.}

\item{colores}{Colores a utilizar en gráficos, en formato lista.}

\item{box}{Graficar o no un Boxplot. Default en TRUE.}

\item{qq}{Graficar o no un QQplot. Default en TRUE.}

\item{shapiro}{Realizar o no un test de Shapiro-Wilks. Default en TRUE}

\item{guardar}{Guardar o no los gráficos y data.frame con medidas de resumen en el Global Environment. Default en FALSE.}
}
\description{
Esta función nos permite realizar descripciones rápidas de variables numéricas, de manera univariada ,bivariada en función de una variable categórica de hasta 3 niveles.
}
\examples{
describir()
}
\keyword{describir}
\keyword{leacs}
