% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_forward.R
\name{step_forward}
\alias{step_forward}
\title{Funci贸n step_forward}
\usage{
step_forward(
  data = data,
  x = "NULL",
  var_list = NULL,
  alpha = 0.2,
  guardar = FALSE
)
}
\arguments{
\item{data}{Base de datos que contiene las variables del modelo.}

\item{x}{Variable dependiente del modelo.}

\item{var_list}{Default NULL, Lista de variables independientes a ingresar al modelo. Si no se ingresa una lista, la funci贸n usa todas las variables de la base de datos.}

\item{alpha}{Default 0.2. Valor de alpha elegido como punto de corte para el ingreso de las variables.}

\item{guardar}{Por defecto FALSE. Si es TRUE se guarda un data frame con los datos.}
}
\description{
Esta funci贸n nos permite extraer el p-valor y el beta de las variables a ingresar
en un modelo de regresi贸n lineal multiple de forma step-forward.
}
\examples{
step_forward()

}
\keyword{leacs}
\keyword{step_forward}
